requirements.txt
Django==4.2.8
#mysqlclient==2.2.0
mysql-connector-python==9.1.0
python-barcode==0.14.0
qrcode==7.4.2
reportlab==4.0.4
WeasyPrint==58.0
Pillow!=10.0.0,!=10.1.0,!=10.2.0,>=10.4.0
python-dotenv==1.0.0
django-crispy-forms==2.5
crispy-bootstrap5==2024.10
django-filter==23.5
django-import-export==3.3.4
openpyxl==3.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
